move files into query_builder directory
