+++ frontend/src/metabase/query_builder/components/ClampedDescription.jsx
