+++ frontend/src/metabase/query_builder/components/view/ViewHeader.jsx
@@ -8,7 +8,7 @@ import Link from "metabase/components/Link";
 import ButtonBar from "metabase/components/ButtonBar";
 import CollectionBadge from "metabase/questions/components/CollectionBadge";
 import LastEditInfoLabel from "metabase/components/LastEditInfoLabel";
-import SavedQuestionHeaderButton from "metabase/questions/components/SavedQuestionHeaderButton";
+import SavedQuestionHeaderButton from "metabase/query_builder/components/SavedQuestionHeaderButton";
 import ViewSection, { ViewHeading, ViewSubHeading } from "./ViewSection";
 import ViewButton from "metabase/query_builder/components/view/ViewButton";
